refactor(project): remove barrel files and changed to named exporting
