+++ src/components/ErrorMessage/ErrorMessage.tsx
@@ -4,8 +4,6 @@ interface ErrorMessageProps  {
   message: string;
 }
 
-const ErrorMessage = ({ message }: ErrorMessageProps) => {
+export const ErrorMessage = ({ message }: ErrorMessageProps) => {
   return <div className="errorMessageContainer">{message}</div>;
 };
-
-export default ErrorMessage;
