--- src/components/index.ts
@@ -1,4 +1,0 @@
-export { default as ErrorMessage } from "./ErrorMessage";
-export { default as Loader } from "./Loader";
-export { default as PokemonItem } from "./PokemonItem";
-export { default as PokemonList } from "./PokemonList";
