+++ src/constants/apiPaths.ts
@@ -1,8 +1,8 @@
-const API = {
+export const API = {
   BASE_URL: "http://localhost:8080/api",
   PATHS: {
     GET_ALL_POKEMON: "/pokemon",
   },
 };
 
-export default API;
+
