+++ src/constants/errorMessage.ts
@@ -1,5 +1,3 @@
-const ERROR_MESSAGE = {
+export const ERROR_MESSAGE = {
   TRY_AGAIN_LATER: "Something went wrong, Please try again later.",
 };
-
-export default ERROR_MESSAGE;
