--- src/constants/index.ts
@@ -1,2 +1,0 @@
-export {default as API} from "./apiPaths"
-export {default as ERROR_MESSAGE} from "./errorMessage"
