--- src/interfaces/index.ts
@@ -1,2 +1,0 @@
-export type { default as Pokemon} from "./pokemon";
-export type { default as Type } from "./type";
